+++ src/main/java/gregtech/api/objects/GT_SidedTexture.java
@@ -79,10 +79,10 @@ public class GT_SidedTexture implements ITexture, IColorModulationContainer {
         IIcon aIcon = mIconContainer[0].getIcon();
 
         float d_16 = 16.0F;
-        float d3 = (float)aIcon.getInterpolatedU(aRenderer.renderMaxX * d_16);
-        float d4 = (float)aIcon.getInterpolatedU(aRenderer.renderMinX * d_16);
-        float d5 = (float)aIcon.getInterpolatedV(aRenderer.renderMinZ * d_16);
-        float d6 = (float)aIcon.getInterpolatedV(aRenderer.renderMaxZ * d_16);
+        float d3 = aIcon.getInterpolatedU(aRenderer.renderMaxX * d_16);
+        float d4 = aIcon.getInterpolatedU(aRenderer.renderMinX * d_16);
+        float d5 = aIcon.getInterpolatedV(aRenderer.renderMinZ * d_16);
+        float d6 = aIcon.getInterpolatedV(aRenderer.renderMaxZ * d_16);
 
         if (aRenderer.renderMinX < 0.0D || aRenderer.renderMaxX > 1.0D) {
             d3 = aIcon.getMaxU();
@@ -108,10 +108,10 @@ public class GT_SidedTexture implements ITexture, IColorModulationContainer {
         if ((aIcon = mIconContainer[0].getOverlayIcon()) != null) {
             Tessellator.instance.setColorRGBA(128, 128, 128, 255);
 
-            d3 = (float)aIcon.getInterpolatedU(aRenderer.renderMaxX * d_16);
-            d4 = (float)aIcon.getInterpolatedU(aRenderer.renderMinX * d_16);
-            d5 = (float)aIcon.getInterpolatedV(aRenderer.renderMinZ * d_16);
-            d6 = (float)aIcon.getInterpolatedV(aRenderer.renderMaxZ * d_16);
+            d3 = aIcon.getInterpolatedU(aRenderer.renderMaxX * d_16);
+            d4 = aIcon.getInterpolatedU(aRenderer.renderMinX * d_16);
+            d5 = aIcon.getInterpolatedV(aRenderer.renderMinZ * d_16);
+            d6 = aIcon.getInterpolatedV(aRenderer.renderMaxZ * d_16);
 
             if (aRenderer.renderMinX < 0.0D || aRenderer.renderMaxX > 1.0D) {
                 d3 = aIcon.getMaxU();
