+++ src/main/java/gregtech/common/tileentities/machines/multi/GT_MetaTileEntity_LargeTurbine_Steam.java
@@ -73,7 +73,7 @@ public class GT_MetaTileEntity_LargeTurbine_Steam extends GT_MetaTileEntity_Larg
     private int useWater(float input) {
         water = water + input;
         int usage = (int) water;
-        water = water - (int) usage;
+        water = water - usage;
         return usage;
     }
